Generate a random data frame from given configuration

Generate a random data frame from given configuration

random_data_frame(configuration, size)

Arguments

configuration

list, a configuration of columns with all arguments required by vector generator passed as sublists of sublist "columns". Column can be also generated with custom function. Pass "custom_column" as column type and function (or function name) as custom_column_generator. Column generator has to accept argument size and return a vector of this size. Third option is to pass an expression that involves existing columns. This can be a simple one, or a call of an existing function.

size

integer, number of rows to generate.

Value

data.frame

Examples

conf <- list(
 columns = list(
   first_column = list(
     type = "string",
     length = 3
   ),
   second_column = list(
     type = "integer",
     max = 10
   ),
   third_column = list(
     type = "calculated",
     formula = "second_column * 2"
   )
 )
)

random_data_frame(conf, size = 10)
#>    first_column second_column third_column
#> 1           OLa             4            8
#> 2           XTF             5           10
#> 3           6Ro             1            2
#> 4           t8M             3            6
#> 5           29S             7           14
#> 6           jXe             9           18
#> 7           KWh             0            0
#> 8           lDg             2            4
#> 9           uB2             2            4
#> 10          ClM             9           18